pyproject.toml
[project]
name = "iaka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.119.1",
    "langchain>=1.0.1",
    "langchain-community>=0.4",
    "langchain-google-genai>=3.0.0",
    "langchain-huggingface>=1.0.0",
    "langgraph>=1.0.1",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.1",
    "unstructured>=0.18.15",
    "uvicorn>=0.38.0",
]
